Adding support for Google "Installed Applications" special redirect URI workflow
Detects the use of "urn:ietf:wg:oauth:2.0:oob" as the redirect URI which
signifies the use of an installed application using the Google OAuth 2
implementation.  With this redirect URI, a custom workflow is used
whereby the authorization code is obtained from the page title as
opposed to the URL fragment in the authorization code grant workflow.